Refactor some code and change player input to method
Reason is there is a bug where a user can input nothing and still the code works making it a draw so turned input into method so unless user wont input string the method won't move forward and keep asking input again and again. while the condition to check input is still not implemented as of now but next step is implementing that condition so the user cannot able to  play around with random strings and empty strings.